requirements.txt
google-cloud-bigquery==3.40.1
python-dotenv==1.2.2
